+++ nym-vpn-core/crates/nym-vpn-lib-types/src/lib.rs
@@ -1,0 +1,16 @@
+// Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+mod connection_data;
+mod tunnel_event;
+mod tunnel_state;
+
+pub use connection_data::{
+    ConnectionData, MixnetConnectionData, TunnelConnectionData, WireguardConnectionData,
+    WireguardNode,
+};
+pub use tunnel_event::{
+    BandwidthEvent, ConnectionEvent, ConnectionStatisticsEvent, MixnetEvent, SphinxPacketRates,
+    TunnelEvent,
+};
+pub use tunnel_state::{ActionAfterDisconnect, ErrorStateReason, TunnelState, TunnelType};
