+++ nym-vpn-core/crates/nym-vpn-lib-types/src/tunnel_state.rs
@@ -1,0 +1,157 @@
+// Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+use std::fmt;
+
+use super::connection_data::{ConnectionData, TunnelConnectionData};
+
+#[derive(Debug, Clone, Copy, Eq, PartialEq)]
+pub enum TunnelType {
+    Mixnet,
+    Wireguard,
+}
+
+/// Public enum describing the tunnel state
+#[derive(Debug, Clone, Eq, PartialEq)]
+pub enum TunnelState {
+    /// Tunnel is disconnected and network connectivity is available.
+    Disconnected,
+
+    /// Tunnel connection is being established.
+    Connecting {
+        connection_data: Option<ConnectionData>,
+    },
+
+    /// Tunnel is connected.
+    Connected { connection_data: ConnectionData },
+
+    /// Tunnel is disconnecting.
+    Disconnecting {
+        after_disconnect: ActionAfterDisconnect,
+    },
+
+    /// Tunnel is disconnected due to failure.
+    Error(ErrorStateReason),
+
+    /// Tunnel is disconnected, network connectivity is unavailable.
+    Offline {
+        /// Whether tunnel will be reconnected upon gaining the network connectivity.
+        reconnect: bool,
+    },
+}
+
+impl fmt::Display for TunnelState {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Disconnected => f.write_str("Disconnected"),
+            Self::Connecting { connection_data } => match connection_data {
+                Some(data) => match data.tunnel {
+                    TunnelConnectionData::Mixnet(ref data) => {
+                        write!(
+                            f,
+                            "Connecting Mixnet tunnel with entry {} and exit {}",
+                            data.nym_address.gateway().to_base58_string(),
+                            data.exit_ipr.gateway().to_base58_string(),
+                        )
+                    }
+                    TunnelConnectionData::Wireguard(ref data) => {
+                        write!(
+                            f,
+                            "Connecting WireGuard tunnel with entry {} and exit {}",
+                            data.entry.endpoint, data.exit.endpoint
+                        )
+                    }
+                },
+                None => f.write_str("Connecting"),
+            },
+            Self::Connected { connection_data } => match connection_data.tunnel {
+                TunnelConnectionData::Mixnet(ref data) => {
+                    write!(
+                        f,
+                        "Connected Mixnet tunnel with entry {} and exit {}",
+                        data.nym_address.gateway().to_base58_string(),
+                        data.exit_ipr.gateway().to_base58_string(),
+                    )
+                }
+                TunnelConnectionData::Wireguard(ref data) => {
+                    write!(
+                        f,
+                        "Connected WireGuard tunnel with entry {} and exit {}",
+                        data.entry.endpoint, data.exit.endpoint
+                    )
+                }
+            },
+            Self::Disconnecting { after_disconnect } => match after_disconnect {
+                ActionAfterDisconnect::Nothing => f.write_str("Disconnecting"),
+                ActionAfterDisconnect::Reconnect => f.write_str("Disconnecting to reconnect"),
+                ActionAfterDisconnect::Error => f.write_str("Disconnecting because of an error"),
+                ActionAfterDisconnect::Offline => {
+                    f.write_str("Disconnecting because device is offline")
+                }
+            },
+            Self::Error(reason) => {
+                write!(f, "Error state: {:?}", reason)
+            }
+            Self::Offline { reconnect } => {
+                if *reconnect {
+                    write!(f, "Offline, auto-connect once back online")
+                } else {
+                    write!(f, "Offline")
+                }
+            }
+        }
+    }
+}
+
+/// Public enum describing action to perform after disconnect
+#[derive(Debug, Clone, Copy, Eq, PartialEq)]
+pub enum ActionAfterDisconnect {
+    /// Do nothing after disconnect
+    Nothing,
+
+    /// Reconnect after disconnect
+    Reconnect,
+
+    /// Enter offline after disconnect
+    Offline,
+
+    /// Enter error state
+    Error,
+}
+
+#[derive(Debug, Clone, Copy, Eq, PartialEq)]
+pub enum ErrorStateReason {
+    /// Issues related to firewall configuration.
+    Firewall,
+
+    /// Failure to configure routing.
+    Routing,
+
+    /// Failure to configure dns.
+    Dns,
+
+    /// Failure to configure tunnel device.
+    TunDevice,
+
+    /// Failure to configure packet tunnel provider.
+    TunnelProvider,
+
+    /// Same entry and exit gateway are unsupported.
+    SameEntryAndExitGateway,
+
+    /// Invalid country set for entry gateway
+    InvalidEntryGatewayCountry,
+
+    /// Invalid country set for exit gateway
+    InvalidExitGatewayCountry,
+
+    /// Gateway is not responding or responding badly to a bandwidth
+    /// increase request, causing credential waste
+    BadBandwidthIncrease,
+
+    /// Failure to duplicate tunnel file descriptor.
+    DuplicateTunFd,
+
+    /// Program errors that must not happen.
+    Internal,
+}
