+++ nym-vpn-core/crates/nym-vpn-proto/Cargo.toml
@@ -22,6 +22,7 @@ nym-vpn-account-controller = { path = "../nym-vpn-account-controller", optional 
 nym-vpn-api-client = { path = "../nym-vpn-api-client", optional = true }
 nym-vpn-network-config = { path = "../nym-vpn-network-config", optional = true }
 nym-vpnd-types = { path = "../nym-vpnd-types", optional = true }
+nym-vpn-lib-types = { path = "../nym-vpn-lib-types", optional = true }
 thiserror = { workspace = true, optional = true }
 time = { workspace = true, optional = true }
 url = { workspace = true, optional = true }
@@ -38,6 +39,7 @@ conversions = [
     "dep:nym-vpn-api-client",
     "dep:nym-vpn-network-config",
     "dep:nym-vpnd-types",
+    "dep:nym-vpn-lib-types",
     "dep:thiserror",
     "dep:time",
     "dep:url",
