+++ nym-vpn-core/crates/nym-vpnd/Cargo.toml
@@ -56,6 +56,7 @@ nym-validator-client.workspace = true
 nym-vpn-account-controller = { path = "../nym-vpn-account-controller" }
 nym-vpn-api-client = { path = "../nym-vpn-api-client" }
 nym-vpn-lib = { path = "../nym-vpn-lib" }
+nym-vpn-lib-types = { path = "../nym-vpn-lib-types" }
 nym-vpn-network-config = { path = "../nym-vpn-network-config" }
 nym-vpn-proto = { path = "../nym-vpn-proto", features = ["conversions"] }
 nym-vpn-store = { path = "../nym-vpn-store" }
